친구 기능 프론트 -> 백엔드 계약서 작성 절차 추가(내 정보 받아넣는거 수정해야 함)
